Metrics tests
