+++ akka-projection-core-test/src/test/scala/akka/projection/StatusObserverSpec.scala
