+++ akka-projection-core-test/src/test/scala/akka/projection/internal/TelemetryProviderSpec.scala
