+++ akka-projection-core-test/src/test/scala/akka/projection/internal/metrics/LifecycleMetricSpec.scala
@@ -12,10 +12,11 @@ import akka.projection.internal.GroupedHandlerStrategy
 import akka.projection.internal.SingleHandlerStrategy
 import akka.projection.internal.metrics.tools.InMemInstrumentsRegistry
 import akka.projection.internal.metrics.tools.InternalProjectionStateMetricsSpec
-import akka.projection.internal.metrics.tools.InternalProjectionStateMetricsSpec._
+import akka.projection.internal.metrics.tools.InternalProjectionStateMetricsSpec.Envelope
+import akka.projection.internal.metrics.tools.InternalProjectionStateMetricsSpec.TelemetryTester
 import akka.projection.internal.metrics.tools.TestHandlers
 
-class WaitTimeMetricSpec extends InternalProjectionStateMetricsSpec {
+class LifecycleMetricSpec extends InternalProjectionStateMetricsSpec {
 
   implicit var projectionId: ProjectionId = null
   before {
@@ -27,51 +28,66 @@ class WaitTimeMetricSpec extends InternalProjectionStateMetricsSpec {
   val defaultNumberOfEnvelopes = 6
 
 }
+class LifecycleMetricAtLeastOnceSpec extends LifecycleMetricSpec {
 
-class WaitTimeMetricAtLeastOnceSpec extends WaitTimeMetricSpec {
-
-  "A metric reporting Wait Time" must {
+  "A metric reporting projection lifecycle metrics" must {
+    // at-least-once
     " in `at-least-once` with singleHandler" must {
-      "report the creation time" in {
+      "count a start and a stop" in {
         val numOfEnvelopes = 20
         val tt: TelemetryTester =
           new TelemetryTester(AtLeastOnce(), SingleHandlerStrategy(TestHandlers.single), numOfEnvelopes)
         runInternal(tt.projectionState) {
-          instruments.creationTimestampInvocations.get should be(numOfEnvelopes)
-          instruments.lastCreationTimestamp.get should be < System.currentTimeMillis()
+          instruments.offsetsSuccessfullyCommitted.get should be(numOfEnvelopes)
+        }
+        eventually {
+          instruments.startedInvocations.get should be(1)
+          instruments.stoppedInvocations.get should be(1)
         }
       }
-      "report the creation time under failing scenarios" in {
+      "count projection failures" in {
         val single = TestHandlers.singleWithErrors(1, 1, 1, 1, 2, 2, 3, 4, 5)
         val tt = new TelemetryTester(
           AtLeastOnce(recoveryStrategy = Some(HandlerRecoveryStrategy.fail)),
           SingleHandlerStrategy(single))
 
         runInternal(tt.projectionState) {
-          instruments.creationTimestampInvocations.get should be >= (1 + 1 + 1 + 1 + 2 + 2 + 3 + 4 + 5 + 6)
-          instruments.lastCreationTimestamp.get should be < System.currentTimeMillis()
+          instruments.offsetsSuccessfullyCommitted.get should be(defaultNumberOfEnvelopes)
+        }
+        eventually {
+          instruments.startedInvocations.get should be(10)
+          instruments.stoppedInvocations.get should be(10)
+          instruments.failureInvocations.get should be(9)
         }
       }
     }
 
     " in `at-least-once` with groupedHandler" must {
-      "report the creation time" in {
+      "report nothing in happy scenarios" in {
         val tt = new TelemetryTester(AtLeastOnce(), GroupedHandlerStrategy(TestHandlers.grouped))
 
         runInternal(tt.projectionState) {
-          instruments.creationTimestampInvocations.get should be(defaultNumberOfEnvelopes)
-          instruments.lastCreationTimestamp.get should be < System.currentTimeMillis()
+          instruments.offsetsSuccessfullyCommitted.get should be(defaultNumberOfEnvelopes)
+        }
+        eventually {
+          instruments.startedInvocations.get should be(1)
+          instruments.stoppedInvocations.get should be(1)
+          instruments.failureInvocations.get should be(0)
         }
       }
     }
     " in `at-least-once` with flowHandler" must {
-      "report the creation time" in {
+      "report nothing in happy scenarios" in {
         val tt =
           new TelemetryTester(AtLeastOnce(), FlowHandlerStrategy[Envelope](TestHandlers.flow))
 
         runInternal(tt.projectionState) {
-          instruments.creationTimestampInvocations.get should be(defaultNumberOfEnvelopes)
-          instruments.lastCreationTimestamp.get should be < System.currentTimeMillis()
+          instruments.offsetsSuccessfullyCommitted.get should be(defaultNumberOfEnvelopes)
+        }
+        eventually {
+          instruments.startedInvocations.get should be(1)
+          instruments.stoppedInvocations.get should be(1)
+          instruments.failureInvocations.get should be(0)
         }
       }
     }
