+++ akka-projection-core-test/src/test/scala/akka/projection/internal/metrics/WaitTimeMetricSpec.scala
@@ -12,10 +12,11 @@ import akka.projection.internal.GroupedHandlerStrategy
 import akka.projection.internal.SingleHandlerStrategy
 import akka.projection.internal.metrics.tools.InMemInstrumentsRegistry
 import akka.projection.internal.metrics.tools.InternalProjectionStateMetricsSpec
-import akka.projection.internal.metrics.tools.InternalProjectionStateMetricsSpec._
+import akka.projection.internal.metrics.tools.InternalProjectionStateMetricsSpec.Envelope
+import akka.projection.internal.metrics.tools.InternalProjectionStateMetricsSpec.TelemetryTester
 import akka.projection.internal.metrics.tools.TestHandlers
 
-class LifecycleMetricSpec extends InternalProjectionStateMetricsSpec {
+class WaitTimeMetricSpec extends InternalProjectionStateMetricsSpec {
 
   implicit var projectionId: ProjectionId = null
   before {
@@ -27,66 +28,51 @@ class LifecycleMetricSpec extends InternalProjectionStateMetricsSpec {
   val defaultNumberOfEnvelopes = 6
 
 }
-class LifecycleMetricAtLeastOnceSpec extends LifecycleMetricSpec {
 
-  "A metric reporting projection lifecycle metrics" must {
-    // at-least-once
+class WaitTimeMetricAtLeastOnceSpec extends WaitTimeMetricSpec {
+
+  "A metric reporting Wait Time" must {
     " in `at-least-once` with singleHandler" must {
-      "count a start and a stop" in {
+      "report the creation time" in {
         val numOfEnvelopes = 20
         val tt: TelemetryTester =
           new TelemetryTester(AtLeastOnce(), SingleHandlerStrategy(TestHandlers.single), numOfEnvelopes)
         runInternal(tt.projectionState) {
-          instruments.offsetsSuccessfullyCommitted.get should be(numOfEnvelopes)
-        }
-        eventually {
-          instruments.startedInvocations.get should be(1)
-          instruments.stoppedInvocations.get should be(1)
+          instruments.creationTimestampInvocations.get should be(numOfEnvelopes)
+          instruments.lastCreationTimestamp.get should be < System.currentTimeMillis()
         }
       }
-      "count projection failures" in {
+      "report the creation time under failing scenarios" in {
         val single = TestHandlers.singleWithErrors(1, 1, 1, 1, 2, 2, 3, 4, 5)
         val tt = new TelemetryTester(
           AtLeastOnce(recoveryStrategy = Some(HandlerRecoveryStrategy.fail)),
           SingleHandlerStrategy(single))
 
         runInternal(tt.projectionState) {
-          instruments.offsetsSuccessfullyCommitted.get should be(defaultNumberOfEnvelopes)
-        }
-        eventually {
-          instruments.startedInvocations.get should be(10)
-          instruments.stoppedInvocations.get should be(10)
-          instruments.failureInvocations.get should be(9)
+          instruments.creationTimestampInvocations.get should be >= (1 + 1 + 1 + 1 + 2 + 2 + 3 + 4 + 5 + 6)
+          instruments.lastCreationTimestamp.get should be < System.currentTimeMillis()
         }
       }
     }
 
     " in `at-least-once` with groupedHandler" must {
-      "report nothing in happy scenarios" in {
+      "report the creation time" in {
         val tt = new TelemetryTester(AtLeastOnce(), GroupedHandlerStrategy(TestHandlers.grouped))
 
         runInternal(tt.projectionState) {
-          instruments.offsetsSuccessfullyCommitted.get should be(defaultNumberOfEnvelopes)
-        }
-        eventually {
-          instruments.startedInvocations.get should be(1)
-          instruments.stoppedInvocations.get should be(1)
-          instruments.failureInvocations.get should be(0)
+          instruments.creationTimestampInvocations.get should be(defaultNumberOfEnvelopes)
+          instruments.lastCreationTimestamp.get should be < System.currentTimeMillis()
         }
       }
     }
     " in `at-least-once` with flowHandler" must {
-      "report nothing in happy scenarios" in {
+      "report the creation time" in {
         val tt =
           new TelemetryTester(AtLeastOnce(), FlowHandlerStrategy[Envelope](TestHandlers.flow))
 
         runInternal(tt.projectionState) {
-          instruments.offsetsSuccessfullyCommitted.get should be(defaultNumberOfEnvelopes)
-        }
-        eventually {
-          instruments.startedInvocations.get should be(1)
-          instruments.stoppedInvocations.get should be(1)
-          instruments.failureInvocations.get should be(0)
+          instruments.creationTimestampInvocations.get should be(defaultNumberOfEnvelopes)
+          instruments.lastCreationTimestamp.get should be < System.currentTimeMillis()
         }
       }
     }
