+++ akka-projection-core-test/src/test/scala/akka/projection/internal/metrics/tools/InMemTelemetry.scala
