+++ akka-projection-core-test/src/test/scala/akka/projection/internal/metrics/tools/TestHandlers.scala
@@ -1,7 +1,3 @@
-/*
- * Copyright (C) 2020 Lightbend Inc. <https://www.lightbend.com>
- */
-
 package akka.projection.internal.metrics.tools
 
 import scala.collection.immutable
